Time: 23 ms (88.03%), Space: 16.8 MB (98.91%) - LeetHub

+++ 0136-single-number/0136-single-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int singleNumber(vector<int>& nums) {
-        int s = 0;
+        long long s = 0;
         for(int x: nums) s^=x;
         return s;
     }
